blockage properties fix
blockage properties fix

+++ qmlsettings_addkeys.cpp
@@ -294,10 +294,10 @@ void QMLSettings::setupKeys() {
     addKey(QString("setDisplay_topTrackNum"),QString("display/topTrackNum"),"bool");
     addKey(QString("setSeed_blockCountMin"),QString("seed/blockCountMin"),"double");
     addKey(QString("setSeed_blockCountMax"),QString("seed/blockCountMax"),"double");
-    addKey(QString("setSeed_blockRow1"),QString("seed/blockRow1"),"int");
-    addKey(QString("setSeed_blockRow2"),QString("seed/blockRow2"),"int");
-    addKey(QString("setSeed_blockRow3"),QString("seed/blockRow3"),"int");
-    addKey(QString("setSeed_blockRow4"),QString("seed/blockRow4"),"int");
-    addKey(QString("setSeed_numRows"),QString("seed/numRows"),"int");
+    addKey(QString("setSeed_blockRow1"),QString("seed/blockRow1"),"double");
+    addKey(QString("setSeed_blockRow2"),QString("seed/blockRow2"),"double");
+    addKey(QString("setSeed_blockRow3"),QString("seed/blockRow3"),"double");
+    addKey(QString("setSeed_blockRow4"),QString("seed/blockRow4"),"double");
+    addKey(QString("setSeed_numRows"),QString("seed/numRows"),"double");
     addKey(QString("setSeed_blockageIsOn"),QString("seed/blockageIsOn"),"bool");
 }
